refactor hubl linting logic
